// app/agentic/components/enums.ts
export enum AgenticHomeViews {
  HOME,
  NEW_SESSION,
}

export enum AgenticNewSessionViews {
  PRIMARY_USE_CASE = 'PRIMARY_USE_CASE',
  ENDPOINTS_SELECTION = 'ENDPOINTS_SELECTION',
  NEW_ENDPOINT_FORM = 'NEW_ENDPOINT_FORM',
  EDIT_ENDPOINT_FORM = 'EDIT_ENDPOINT_FORM',
  COOKBOOKS_SELECTION = 'COOKBOOKS_SELECTION',
  AGENTIC_RUN_FORM = 'AGENTIC_RUN_FORM',
}